pyproject.toml
[project]
name = "pegasus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "brightdata>=0.4.0.4",
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.0",
    "langchain>=1.0",
    "langchain-core>=0.3",
    "langchain-google-genai>=4.0",
    "langgraph>=0.3",
    "sse-starlette>=2.0",
    "python-dotenv>=1.0",
    "redis>=5.0.0",
    "uvicorn>=0.41.0",
]
